+++ Espera/Espera.Core/Management/Library.cs
@@ -29,7 +29,6 @@ namespace Espera.Core.Management
 
         private readonly AccessControl accessControl;
         private readonly AudioPlayer audioPlayer;
-        private readonly Subject<Playlist> currentPlaylistChanged;
         private readonly IFileSystem fileSystem;
         private readonly CompositeDisposable globalSubscriptions;
         private readonly BehaviorSubject<bool> isUpdating;
@@ -45,6 +44,7 @@ namespace Espera.Core.Management
         private readonly Subject<Unit> songsUpdated;
         private readonly ObservableAsPropertyHelper<float> volume;
         private Playlist currentPlayingPlaylist;
+        private Playlist currentPlaylist;
         private IDisposable currentSongFinderSubscription;
         private DateTime lastSongAddTime;
 
@@ -62,9 +62,6 @@ namespace Espera.Core.Management
             this.songLock = new ReaderWriterLockSlim();
             this.songs = new HashSet<LocalSong>();
             this.playlists = new ReactiveList<Playlist>();
-            this.currentPlaylistChanged = new Subject<Playlist>();
-            this.CanPlayNextSong = this.currentPlaylistChanged.Select(x => x.CanPlayNextSong).Switch();
-            this.CanPlayPreviousSong = this.currentPlaylistChanged.Select(x => x.CanPlayPreviousSong).Switch();
             this.songSourcePath = new BehaviorSubject<string>(null);
             this.songsUpdated = new Subject<Unit>();
             this.audioPlayer = new AudioPlayer();
@@ -75,7 +72,7 @@ namespace Espera.Core.Management
             this.TotalTime = this.audioPlayer.TotalTime;
             this.PlaybackState = this.audioPlayer.PlaybackState;
 
-            this.CanPlayNextSong.SampleAndCombineLatest(this.audioPlayer.PlaybackState
+            this.WhenAnyValue(x => x.CurrentPlaylist.CanPlayNextSong).SampleAndCombineLatest(this.audioPlayer.PlaybackState
                     .Where(p => p == AudioPlayerState.Finished), (canPlayNextSong, _) => canPlayNextSong)
                 .SelectMany(x => this.HandleSongFinishAsync(x).ToObservable())
                 .Subscribe();
@@ -86,25 +83,12 @@ namespace Espera.Core.Management
                 .ToProperty(this, x => x.Volume);
         }
 
-        /// <summary>
-        /// Gets a value indicating whether the next song in the playlist can be played.
-        /// </summary>
-        /// <value>true if the next song in the playlist can be played; otherwise, false.</value>
-        public IObservable<bool> CanPlayNextSong { get; private set; }
-
-        /// <summary>
-        /// Gets a value indicating whether the previous song in the playlist can be played.
-        /// </summary>
-        /// <value>true if the previous song in the playlist can be played; otherwise, false.</value>
-        public IObservable<bool> CanPlayPreviousSong { get; private set; }
-
         public IObservable<TimeSpan> CurrentPlaybackTime { get; private set; }
 
-        public Playlist CurrentPlaylist { get; private set; }
-
-        public IObservable<Playlist> CurrentPlaylistChanged
+        public Playlist CurrentPlaylist
         {
-            get { return this.currentPlaylistChanged.AsObservable(); }
+            get { return this.currentPlaylist; }
+            set { this.RaiseAndSetIfChanged(ref this.currentPlaylist, value); }
         }
 
         /// <summary>
@@ -129,21 +113,12 @@ namespace Espera.Core.Management
         public IObservable<AudioPlayerState> PlaybackState { get; private set; }
 
         /// <summary>
-        /// Returns an enumeration of playlists that implements <see cref="INotifyCollectionChanged" />.
+        /// Returns an enumeration of playlists that implements
+        /// <see cref="INotifyCollectionChanged" /> .
         /// </summary>
         public IReadOnlyReactiveList<Playlist> Playlists
         {
             get { return this.playlists; }
-        }
-
-        public TimeSpan RemainingPlaylistTimeout
-        {
-            get
-            {
-                return this.lastSongAddTime + this.settings.PlaylistTimeout <= DateTime.Now
-                           ? TimeSpan.Zero
-                           : this.lastSongAddTime - DateTime.Now + this.settings.PlaylistTimeout;
-            }
         }
 
         public IRemoteAccessControl RemoteAccessControl
@@ -209,6 +184,37 @@ namespace Espera.Core.Management
         }
 
         /// <summary>
+        /// Adds the specified song to the end of the playlist as guest.
+        /// </summary>
+        /// <remarks>
+        /// <para>This method is intended only for guest access tokens.</para>
+        /// <para>
+        /// As soon as the song is added to the playlist, the entry is marked as "shadow voted".
+        /// This means that it won't be favoured over other songs, like a regular vote, but stays at
+        /// the end of the playlist.
+        /// </para>
+        /// <para>
+        /// Shadow votes still decrease the available votes of the guest like regular votes, this
+        /// prevents guests from spamming songs to the playlist.
+        /// </para>
+        /// </remarks>
+        /// <param name="song">The song to add to the end of the playlist.</param>
+        /// <param name="accessToken">The access token. Must have guest permission.</param>
+        /// <exception cref="InvalidOperationException">The guest system is disabled.</exception>
+        /// <exception cref="AccessException">The access token isn't a guest token.</exception>
+        public void AddGuestSongToPlaylist(Song song, Guid accessToken)
+        {
+            if (song == null)
+                Throw.ArgumentNullException(() => song);
+
+            this.accessControl.VerifyVotingPreconditions(accessToken);
+
+            PlaylistEntry entry = this.CurrentPlaylist.AddShadowVotedSong(song);
+
+            this.accessControl.RegisterShadowVote(accessToken, entry);
+        }
+
+        /// <summary>
         /// Adds a new playlist with the specified name to the library.
         /// </summary>
         /// <param name="name">
@@ -243,28 +249,6 @@ namespace Espera.Core.Management
             this.accessControl.VerifyAccess(accessToken);
 
             this.CurrentPlaylist.AddSongs(songList.ToList()); // Copy the sequence to a list, so that the enumeration doesn't gets modified
-        }
-
-        /// <summary>
-        /// Adds the song to the end of the playlist. This method throws an exception, if there is
-        /// an outstanding timeout.
-        /// </summary>
-        /// <param name="song">The song to add to the end of the playlist.</param>
-        /// <exception cref="InvalidOperationException">There is an outstanding playlist timeout.</exception>
-        public void AddSongToPlaylist(Song song)
-        {
-            if (song == null)
-                Throw.ArgumentNullException(() => song);
-
-            if (this.settings.EnablePlaylistTimeout && this.RemainingPlaylistTimeout > TimeSpan.Zero)
-                throw new InvalidOperationException("Current playlist has a remaining timeout.");
-
-            this.CurrentPlaylist.AddSongs(new[] { song });
-
-            if (this.settings.EnablePlaylistTimeout)
-            {
-                this.lastSongAddTime = DateTime.Now;
-            }
         }
 
         public void ChangeSongSourcePath(string path, Guid accessToken)
@@ -398,7 +382,7 @@ namespace Espera.Core.Management
         /// </summary>
         public async Task PlayPreviousSongAsync(Guid accessToken)
         {
-            this.accessControl.VerifyAccess(accessToken, await this.CurrentPlaylist.CanPlayPreviousSong.FirstAsync());
+            this.accessControl.VerifyAccess(accessToken, this.CurrentPlaylist.CanPlayPreviousSong);
 
             if (!this.CurrentPlaylist.CurrentSongIndex.HasValue)
                 throw new InvalidOperationException("The previous song can't be played as there is no current playlist index.");
@@ -537,7 +521,6 @@ namespace Espera.Core.Management
             this.accessControl.VerifyAccess(accessToken, this.settings.LockPlaylist);
 
             this.CurrentPlaylist = playlist;
-            this.currentPlaylistChanged.OnNext(playlist);
         }
 
         /// <summary>
@@ -559,7 +542,7 @@ namespace Espera.Core.Management
 
         private async Task HandleSongCorruptionAsync()
         {
-            if (!await this.CurrentPlaylist.CanPlayNextSong.FirstAsync())
+            if (!this.CurrentPlaylist.CanPlayNextSong)
             {
                 this.CurrentPlaylist.CurrentSongIndex = null;
             }
@@ -583,14 +566,14 @@ namespace Espera.Core.Management
             }
         }
 
-        private async Task InternPlayNextSongAsync()
+        private Task InternPlayNextSongAsync()
         {
-            if (!await this.CurrentPlaylist.CanPlayNextSong.FirstAsync() || !this.CurrentPlaylist.CurrentSongIndex.HasValue)
+            if (!this.CurrentPlaylist.CanPlayNextSong || !this.CurrentPlaylist.CurrentSongIndex.HasValue)
                 throw new InvalidOperationException("The next song couldn't be played.");
 
             int nextIndex = this.CurrentPlaylist.CurrentSongIndex.Value + 1;
 
-            await this.InternPlaySongAsync(nextIndex);
+            return this.InternPlaySongAsync(nextIndex);
         }
 
         private async Task InternPlaySongAsync(int playlistIndex)
@@ -891,6 +874,8 @@ namespace Espera.Core.Management
                     }
                 }, () =>
                 {
+                    this.Save();
+
                     this.StartOnlineArtworkLookup();
 
                     this.songLock.EnterReadLock();