+++ Espera/Espera.View/ViewModels/ISongSourceViewModel.cs
@@ -23,7 +23,5 @@ namespace Espera.View.ViewModels
         string SearchText { get; set; }
 
         IEnumerable<ISongViewModelBase> SelectedSongs { get; set; }
-
-        IObservable<Unit> TimeoutWarning { get; }
     }
 }