+++ Espera/Espera.View/ViewModels/SettingsViewModel.cs
@@ -11,6 +11,7 @@ using Espera.Core;
 using Espera.Core.Management;
 using Espera.Core.Mobile;
 using Espera.Core.Settings;
+using Rareform.Reflection;
 using Rareform.Validation;
 using ReactiveUI;
 using Splat;
@@ -40,8 +41,10 @@ namespace Espera.View.ViewModels
 
         private readonly ObservableAsPropertyHelper<bool> enableChangelog;
         private readonly ObservableAsPropertyHelper<bool> isPortOccupied;
+        private readonly ObservableAsPropertyHelper<bool> isRemoteAccessReallyLocked;
         private readonly Library library;
         private readonly ObservableAsPropertyHelper<string> librarySource;
+        private readonly ObservableAsPropertyHelper<bool> showRemoteControlPasswordError;
         private readonly ViewSettings viewSettings;
         private readonly IWindowManager windowManager;
         private string creationPassword;
@@ -154,6 +157,12 @@ namespace Espera.View.ViewModels
                 .Select(x => !String.IsNullOrWhiteSpace(x)));
             this.ChangeRemoteControlPasswordCommand.Subscribe(x =>
                 this.library.RemoteAccessControl.SetRemotePassword(this.accessToken, this.RemoteControlPassword));
+            this.showRemoteControlPasswordError = this.WhenAnyValue(x => x.RemoteControlPassword, x => x.LockRemoteControl,
+                    (password, lockRemoteControl) => String.IsNullOrWhiteSpace(password) && lockRemoteControl)
+                .ToProperty(this, x => x.ShowRemoteControlPasswordError);
+
+            this.isRemoteAccessReallyLocked = this.library.RemoteAccessControl.WhenAnyValue(x => x.IsRemoteAccessReallyLocked)
+                .ToProperty(this, x => x.IsRemoteAccessReallyLocked);
 
             this.isPortOccupied = mobileApiInfo.IsPortOccupied.ToProperty(this, x => x.IsPortOccupied);
 
@@ -274,14 +283,10 @@ namespace Espera.View.ViewModels
             set { this.viewSettings.EnableChangelog = value; }
         }
 
-        public bool EnablePlaylistTimeout
+        public bool EnableGuestSystem
         {
-            get { return this.coreSettings.EnablePlaylistTimeout; }
-            set
-            {
-                this.coreSettings.EnablePlaylistTimeout = value;
-                this.RaisePropertyChanged();
-            }
+            get { return this.coreSettings.EnableGuestSystem; }
+            set { this.coreSettings.EnableGuestSystem = value; }
         }
 
         public bool EnableRemoteControl
@@ -294,11 +299,7 @@ namespace Espera.View.ViewModels
             }
         }
 
-        public bool EnableVotingSystem
-        {
-            get { return this.coreSettings.EnableVotingSystem; }
-            set { this.coreSettings.EnableVotingSystem = value; }
-        }
+        public string Error { get; private set; }
 
         public bool GoFullScreenOnLock
         {
@@ -314,6 +315,11 @@ namespace Espera.View.ViewModels
         public bool IsPortOccupied
         {
             get { return this.isPortOccupied.Value; }
+        }
+
+        public bool IsRemoteAccessReallyLocked
+        {
+            get { return this.isRemoteAccessReallyLocked.Value; }
         }
 
         public bool IsWrongPassword
@@ -384,10 +390,9 @@ namespace Espera.View.ViewModels
 
         public ReactiveCommand<object> OpenLinkCommand { get; private set; }
 
-        public int PlaylistTimeout
+        public string PlayStoreLink
         {
-            get { return (int)this.coreSettings.PlaylistTimeout.TotalSeconds; }
-            set { this.coreSettings.PlaylistTimeout = TimeSpan.FromSeconds(value); }
+            get { return "http://play.google.com/store/apps/details?id=com.flagbug.esperamobile"; }
         }
 
         public int Port
@@ -423,6 +428,11 @@ namespace Espera.View.ViewModels
         {
             get { return this.showLogin; }
             set { this.RaiseAndSetIfChanged(ref this.showLogin, value); }
+        }
+
+        public bool ShowRemoteControlPasswordError
+        {
+            get { return this.showRemoteControlPasswordError.Value; }
         }
 
         public bool ShowSettings