+++ Espera/Espera.View/ViewModels/ShellViewModel.cs
@@ -29,12 +29,10 @@ namespace Espera.View.ViewModels
         private readonly ObservableAsPropertyHelper<ISongSourceViewModel> currentSongSource;
         private readonly ObservableAsPropertyHelper<string> currentTime;
         private readonly ObservableAsPropertyHelper<IReactiveCommand> defaultPlaybackCommand;
-        private readonly ObservableAsPropertyHelper<bool> displayTimeoutWarning;
         private readonly CompositeDisposable disposable;
         private readonly ObservableAsPropertyHelper<bool> isAdmin;
         private readonly ObservableAsPropertyHelper<bool> isPlaying;
         private readonly Library library;
-        private readonly ObservableAsPropertyHelper<bool> showPlaylistTimeout;
         private readonly ObservableAsPropertyHelper<bool> showVotes;
         private readonly ObservableAsPropertyHelper<int> totalSeconds;
         private readonly ObservableAsPropertyHelper<string> totalTime;
@@ -57,7 +55,7 @@ namespace Espera.View.ViewModels
             this.library.Initialize();
             this.accessToken = this.library.LocalAccessControl.RegisterLocalAccessToken();
 
-            this.library.CurrentPlaylistChanged.Subscribe(x => this.RaisePropertyChanged("CurrentPlaylist"));
+            this.library.WhenAnyValue(x => x.CurrentPlaylist).Subscribe(x => this.RaisePropertyChanged("CurrentPlaylist"));
 
             this.canChangeTime = this.HasAccess(this.coreSettings.WhenAnyValue(x => x.LockTime))
                 .ToProperty(this, x => x.CanChangeTime);
@@ -66,8 +64,8 @@ namespace Espera.View.ViewModels
             this.canAlterPlaylist = this.HasAccess(this.coreSettings.WhenAnyValue(x => x.LockPlaylist))
                 .ToProperty(this, x => x.CanAlterPlaylist);
 
-            this.showVotes = this.coreSettings.WhenAnyValue(x => x.EnableVotingSystem)
-                .CombineLatest(mobileApiInfo.ConnectedClientCount, (enableVoting, connectedClients) => enableVoting && connectedClients > 0)
+            this.showVotes = this.library.RemoteAccessControl.WhenAnyValue(x => x.IsGuestSystemReallyEnabled)
+                .CombineLatest(mobileApiInfo.ConnectedClientCount, (enableGuestSystem, connectedClients) => enableGuestSystem && connectedClients > 0)
                 .ToProperty(this, x => x.ShowVotes);
 
             this.isAdmin = this.library.LocalAccessControl.ObserveAccessPermission(this.accessToken)
@@ -75,11 +73,11 @@ namespace Espera.View.ViewModels
                 .ToProperty(this, x => x.IsAdmin);
 
             this.NextSongCommand = ReactiveCommand.CreateAsyncTask(this.HasAccess(this.coreSettings.WhenAnyValue(x => x.LockPlayPause))
-                    .CombineLatest(this.library.CanPlayNextSong, (x1, x2) => x1 && x2),
+                    .CombineLatest(this.library.WhenAnyValue(x => x.CurrentPlaylist.CanPlayNextSong), (x1, x2) => x1 && x2),
                 _ => this.library.PlayNextSongAsync(this.accessToken));
 
             this.PreviousSongCommand = ReactiveCommand.CreateAsyncTask(this.HasAccess(this.coreSettings.WhenAnyValue(x => x.LockPlayPause))
-                    .CombineLatest(this.library.CanPlayPreviousSong, (x1, x2) => x1 && x2),
+                    .CombineLatest(this.library.WhenAnyValue(x => x.CurrentPlaylist.CanPlayPreviousSong), (x1, x2) => x1 && x2),
                 _ => this.library.PlayPreviousSongAsync(this.accessToken));
 
             if (!this.library.Playlists.Any())
@@ -98,11 +96,6 @@ namespace Espera.View.ViewModels
             this.YoutubeViewModel = new YoutubeViewModel(this.library, this.ViewSettings, this.coreSettings, accessToken);
             this.SoundCloudViewModel = new SoundCloudViewModel(this.library, accessToken, this.coreSettings, this.ViewSettings);
             this.DirectYoutubeViewModel = new DirectYoutubeViewModel(this.library, accessToken);
-
-            Observable.Interval(TimeSpan.FromMilliseconds(300), RxApp.TaskpoolScheduler)
-                .Where(_ => this.RemainingPlaylistTimeout > TimeSpan.Zero)
-                .Subscribe(x => this.RaisePropertyChanged("RemainingPlaylistTimeout"))
-                .DisposeWith(this.disposable);
 
             this.currentSongSource = this.WhenAnyValue(x => x.IsLocal, x => x.IsYoutube, x => x.IsSoundCloud,
                 (local, youtube, soundcloud) =>
@@ -125,14 +118,6 @@ namespace Espera.View.ViewModels
                     return this.LocalViewModel;
                 })
                 .ToProperty(this, x => x.CurrentSongSource, null, ImmediateScheduler.Instance);
-
-            this.displayTimeoutWarning = Observable.Merge(this.LocalViewModel.TimeoutWarning, this.YoutubeViewModel.TimeoutWarning, this.DirectYoutubeViewModel.TimeoutWarning, this.SoundCloudViewModel.TimeoutWarning)
-                .SelectMany(x => new[] { true, false }.ToObservable())
-                .ToProperty(this, x => x.DisplayTimeoutWarning);
-
-            this.showPlaylistTimeout = this.WhenAnyValue(x => x.IsAdmin)
-                .CombineLatest(this.WhenAnyValue(x => x.SettingsViewModel.EnablePlaylistTimeout), (isAdmin, enableTimeout) => !isAdmin && enableTimeout)
-                .ToProperty(this, x => x.ShowPlaylistTimeout);
 
             this.MuteCommand = ReactiveCommand.Create(this.WhenAnyValue(x => x.IsAdmin));
             this.MuteCommand.Subscribe(x => this.Volume = 0);
@@ -359,11 +344,6 @@ namespace Espera.View.ViewModels
 
         public DirectYoutubeViewModel DirectYoutubeViewModel { get; private set; }
 
-        public bool DisplayTimeoutWarning
-        {
-            get { return this.displayTimeoutWarning.Value; }
-        }
-
         public ReactiveCommand<object> EditPlaylistNameCommand { get; private set; }
 
         public bool IsAdmin
@@ -439,11 +419,6 @@ namespace Espera.View.ViewModels
         /// </summary>
         public ReactiveCommand<Unit> PreviousSongCommand { get; private set; }
 
-        public TimeSpan RemainingPlaylistTimeout
-        {
-            get { return this.library.RemainingPlaylistTimeout; }
-        }
-
         public ReactiveCommand<object> RemovePlaylistCommand { get; private set; }
 
         public ReactiveCommand<object> RemoveSelectedPlaylistEntriesCommand { get; private set; }
@@ -455,11 +430,6 @@ namespace Espera.View.ViewModels
         }
 
         public SettingsViewModel SettingsViewModel { get; private set; }
-
-        public bool ShowPlaylistTimeout
-        {
-            get { return this.showPlaylistTimeout.Value; }
-        }
 
         public ReactiveCommand<object> ShowSettingsCommand { get; private set; }